moved ValidateIngressForClass to lib and added a check in ingress status delete
