+++ internal/nodes/avi_model_routeingr_hostname_shard.go
@@ -143,7 +143,7 @@ func GetK8sIngressModel(name, namespace, key string) (*K8sIngressModel, error, b
 	if err != nil {
 		return &ingrModel, err, processObj
 	}
-	processObj = validateIngressForClass(key, ingObj) && utils.CheckIfNamespaceAccepted(namespace)
+	processObj = lib.ValidateIngressForClass(key, ingObj) && utils.CheckIfNamespaceAccepted(namespace)
 	ingrModel.spec = ingObj.Spec
 	ingrModel.annotations = ingObj.GetAnnotations()
 	return &ingrModel, nil, processObj
