+++ internal/nodes/ingress_model_rel.go
@@ -32,7 +32,6 @@ import (
 	networkingv1beta1 "k8s.io/api/networking/v1beta1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	servicesapi "sigs.k8s.io/service-apis/apis/v1alpha1"
 	svcapiv1alpha1 "sigs.k8s.io/service-apis/apis/v1alpha1"
 )
@@ -309,7 +308,7 @@ func IngressChanges(ingName string, namespace string, key string) ([]string, boo
 
 		// If the Ingress Class is not found or is not valid, then return.
 		// When the correct Ingress Class is added, then the Ingress would be processed again.
-		if !validateIngressForClass(key, ingObj) {
+		if !lib.ValidateIngressForClass(key, ingObj) {
 			svcToDel := objects.SharedSvcLister().IngressMappings(namespace).RemoveIngressMappings(ingName)
 			if lib.AutoAnnotateNPLSvc() {
 				for _, svc := range svcToDel {
@@ -322,7 +321,7 @@ func IngressChanges(ingName string, namespace string, key string) ([]string, boo
 		if utils.GetIngressClassEnabled() {
 			if ingObj.Spec.IngressClassName != nil {
 				objects.SharedSvcLister().IngressMappings(metav1.NamespaceAll).UpdateIngressClassMappings(namespace+"/"+ingName, *ingObj.Spec.IngressClassName)
-			} else if aviIngClassName, found := isAviLBDefaultIngressClass(); found {
+			} else if aviIngClassName, found := lib.IsAviLBDefaultIngressClass(); found {
 				// check if default IngressClass is present, and it is Avi's IngressClass, in which case add mapping for that.
 				objects.SharedSvcLister().IngressMappings(metav1.NamespaceAll).UpdateIngressClassMappings(namespace+"/"+ingName, aviIngClassName)
 			} else {
@@ -836,78 +835,4 @@ func validateSvcApiGatewayForClass(key string, gateway *svcapiv1alpha1.Gateway) 
 	}
 
 	return nil
-}
-
-func filterIngressOnClassAnnotation(key string, ingress *networkingv1beta1.Ingress) bool {
-	// If Avi is not the default ingress, then filter on ingress class.
-	if !lib.GetDefaultIngController() {
-		annotations := ingress.GetAnnotations()
-		ingClass, ok := annotations[lib.INGRESS_CLASS_ANNOT]
-		if ok && ingClass == lib.AVI_INGRESS_CLASS {
-			return true
-		} else {
-			utils.AviLog.Infof("key: %s, msg: AKO is not running as the default ingress controller. Not processing the ingress: %s. Please annotate the ingress class as 'avi'", key, ingress.Name)
-			return false
-		}
-	} else {
-		// If Avi is the default ingress controller, sync everything than the ones that are annotated with ingress class other than 'avi'
-		annotations := ingress.GetAnnotations()
-		ingClass, ok := annotations[lib.INGRESS_CLASS_ANNOT]
-		if ok && ingClass != lib.AVI_INGRESS_CLASS {
-			utils.AviLog.Infof("key: %s, msg: AKO is the default ingress controller but not processing the ingress: %s since ingress class is set to : %s", key, ingress.Name, ingClass)
-			return false
-		} else {
-			return true
-		}
-	}
-}
-
-func validateIngressForClass(key string, ingress *networkingv1beta1.Ingress) bool {
-	// see whether ingress class resources are present or not
-	if !utils.GetIngressClassEnabled() {
-		return filterIngressOnClassAnnotation(key, ingress)
-	}
-
-	if ingress.Spec.IngressClassName == nil {
-		// check whether avi-lb ingress class is set as the default ingress class
-		if _, found := isAviLBDefaultIngressClass(); found {
-			utils.AviLog.Infof("key: %s, msg: ingress class name is not specified but ako.vmware.com/avi-lb is default ingress controller", key)
-			return true
-		} else {
-			utils.AviLog.Warnf("key: %s, msg: ingress class name not specified for ingress %s and ako.vmware.com/avi-lb is not default ingress controller", key, ingress.Name)
-			return false
-		}
-	}
-
-	ingClassObj, err := utils.GetInformers().IngressClassInformer.Lister().Get(*ingress.Spec.IngressClassName)
-	if err != nil {
-		utils.AviLog.Warnf("key: %s, msg: Unable to fetch corresponding networking.k8s.io/ingressclass %s %v",
-			key, *ingress.Spec.IngressClassName, err)
-		return false
-	}
-
-	// Additional check to see if the gatewayclass is a valid avi gateway class or not.
-	if ingClassObj.Spec.Controller != lib.AviIngressController {
-		// Return an error since this is not our object.
-		utils.AviLog.Warnf("key: %s, msg: Unexpected controller in ingress class %s", key, *ingress.Spec.IngressClassName)
-		return false
-	}
-
-	return true
-}
-
-func isAviLBDefaultIngressClass() (string, bool) {
-	ingClassObjs, _ := utils.GetInformers().IngressClassInformer.Lister().List(labels.Set(nil).AsSelector())
-	for _, ingClass := range ingClassObjs {
-		if ingClass.Spec.Controller == lib.AviIngressController {
-			annotations := ingClass.GetAnnotations()
-			isDefaultClass, ok := annotations[lib.DefaultIngressClassAnnotation]
-			if ok && isDefaultClass == "true" {
-				return ingClass.Name, true
-			}
-		}
-	}
-
-	utils.AviLog.Debugf("IngressClass with controller ako.vmware.com/avi-lb not found in the cluster")
-	return "", false
 }
